Fixed the NN train and test function
Fixed the NN train and test function.
We were doing forward passes on the test set in every iteration of the training loop.
This underfits the data and leads to poor performance.
Instead modified the loop to be able to do that only after the training loop is done.
Also - need to store the CV metrics instead of doing 2 test_step calls.